+++ tests/unit/moduleapi/testrdb.tcl
@@ -62,47 +62,66 @@ tags "modules" {
                     $master config set repl-diskless-sync yes
                     $master config set rdbcompression no
                     $replica config set repl-diskless-load swapdb
+                    $master config set hz 500
+                    $replica config set hz 500
+                    $master config set dynamic-hz no
+                    $replica config set dynamic-hz no
+                    set start [clock clicks -milliseconds]
                     for {set k 0} {$k < 30} {incr k} {
                         r testrdb.set.key key$k [string repeat A [expr {int(rand()*1000000)}]]
                     }
 
+                    if {$::verbose} {
+                        set end [clock clicks -milliseconds]
+                        set duration [expr $end - $start]
+                        puts "filling took $duration ms (TODO: use pipeline)"
+                        set start [clock clicks -milliseconds]
+                    }
+
                     # Start the replication process...
+                    set loglines [count_log_lines -1]
                     $master config set repl-diskless-sync-delay 0
                     $replica replicaof $master_host $master_port
 
                     # kill the replication at various points
-                    set attempts 3
-                    if {$::accurate} { set attempts 10 }
+                    set attempts 100
+                    if {$::accurate} { set attempts 500 }
                     for {set i 0} {$i < $attempts} {incr i} {
                         # wait for the replica to start reading the rdb
                         # using the log file since the replica only responds to INFO once in 2mb
-                        wait_for_log_message -1 "*Loading DB in memory*" 5 2000 1
+                        set res [wait_for_log_messages -1 {"*Loading DB in memory*"} $loglines 2000 1]
+                        set loglines [lindex $res 1]
 
                         # add some additional random sleep so that we kill the master on a different place each time
-                        after [expr {int(rand()*100)}]
+                        after [expr {int(rand()*50)}]
 
                         # kill the replica connection on the master
                         set killed [$master client kill type replica]
 
-                        if {[catch {
-                            set res [wait_for_log_message -1 "*Internal error in RDB*" 5 100 10]
-                            if {$::verbose} {
-                                puts $res
-                            }
-                        }]} {
-                            puts "failed triggering short read"
+                        set res [wait_for_log_messages -1 {"*Internal error in RDB*" "*Finished with success*" "*Successful partial resynchronization*"} $loglines 1000 1]
+                        if {$::verbose} { puts $res }
+                        set log_text [lindex $res 0]
+                        set loglines [lindex $res 1]
+                        if {![string match "*Internal error in RDB*" $log_text]} {
                             # force the replica to try another full sync
+                            $master multi
                             $master client kill type replica
                             $master set asdf asdf
                             # the side effect of resizing the backlog is that it is flushed (16k is the min size)
                             $master config set repl-backlog-size [expr {16384 + $i}]
+                            $master exec
                         }
                         # wait for loading to stop (fail)
-                        wait_for_condition 100 10 {
+                        wait_for_condition 1000 1 {
                             [s -1 loading] eq 0
                         } else {
                             fail "Replica didn't disconnect"
                         }
+                    }
+                    if {$::verbose} {
+                        set end [clock clicks -milliseconds]
+                        set duration [expr $end - $start]
+                        puts "test took $duration ms"
                     }
                     # enable fast shutdown
                     $master config set rdb-key-save-delay 0
